+++ sources/PLoader/src/Utils/Math.cpp
@@ -82,8 +82,8 @@ float Math::RoundFloat(float value, int numDigits)
 
     if (digsInInt < numDigits)  // Подстрахуемся
     {
-        uint pow = Math::Pow10(numDigits - digsInInt);
-        absValue = static_cast<int>(absValue * pow + 0.5F) / static_cast<float>(pow);
+        float pow = static_cast<float>(Math::Pow10(numDigits - digsInInt));
+        absValue = static_cast<float>(static_cast<int>(absValue * pow + 0.5F)) / static_cast<float>(pow);
     }
 
     return (value > 0.0F) ? absValue : -absValue;
